feat(vector store): set up vault to vector store config and commands
Users can save to vector store on save, on app load or never. They also have the possibility of
running vault -> vector store management commands to populate, delete and garbage collect. All bulk
vault save operations use the latest modified time for a file in the vector store as the starting
point to fetch new files to be embedded and stored. This helps save on embeddings costs.